renamed all budgetX to categoryX
