+++ lab-derek/src/actions/category-actions.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import uuid from 'uuid/v1';
 
-export const budgetCreate = category => {
+export const categoryCreate = category => {
   category.id = uuid();
   category.timeStamp = new Date();
   return {
@@ -10,14 +10,14 @@ export const budgetCreate = category => {
   };
 };
 
-export const budgetUpdate = category => {
+export const categoryUpdate = category => {
   return {
     type: 'CATEGORY_UPDATE',
     payload: category,
   };
 };
 
-export const budgetDelete = category => {
+export const categoryDelete = category => {
   return {
     type: 'CATEGORY_DELETE',
     payload: category,
