+++ lab-derek/src/reducer/index.js
@@ -1,6 +1,6 @@
 import {combineReducers} from 'redux';
 
-import categoriesReducer from './budget-category.js';
+import categoriesReducer from './category.js';
 import expensesReducer from './expense.js';
 
 export default combineReducers({
